requirements.txt
anthropic>=0.79.0
httpx>=0.28.0
praw>=7.8.0
playwright>=1.49.0
python-dotenv>=1.0.0
pytest>=8.0.0
